Remove max allele frequency field, create util.js and move isEmptyField utility function there
